requirements.txt
numpy>=1.21.0
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
openpyxl>=3.0.10
pandas>=1.3.0
Pillow>=8.2.0
pyttsx3>=2.90
